Update: Major changes. Add tests. Define Basic Structure.
Major changes. Add GoLangCI-Lint for tests. Update makefile.

Use Cobra as the base for CLI-based interaction. Use `logrus` for
logging.

Add a flag to turn logging on/off. By default, logs will be written
to a log file as well as `stderr`.

Define root command, and the basic interaction required for the
root command. This root command will be the actual program, with
its behaviour being modified depending on the flags/arguments used.

Add test cases. Use monkey-patching for tests to ensure isolation
of methods/flow-control logic when needed.