Update: Major refactor. Store input more graciously
This update includes a major refactor in the codebase. Code from
the previous commit has been broken down into three different
files and grouped on the basis of what each segment is supposed
to achieve.

To simplify testing, replace `rootCommand` with a generator
function and a variable of the same name. The generator function
ensures that a new instance of the `rootCommand` variable can be
generated when required during tests.

Changes to the codebase;
 - The file `root.go` contains code used to define the root command
this includes defining the version, title, and anything else that is
an intrinsic part of the root command object.
 - The file `userInput.go` contains a singular structure that will
hold data from user input/flags and use this data while performing
tasks internally.
 - The file `entryPoint.go`, as the name states, is the main entry
point for the flow-of-control to use the package. It exposes a single
public function `Execute` which then fetches user input and acts
upon.

Add test cases pertaining to the changes made to the code base.
Update documentation, fix minor spelling mistakes and reduce dumb
comments.

Modify logger. Instead of having the logger log to two streams (log
file and stderr), the logger will now push to only one of them. By
default, logs will be written to the log file, in case the log file
cannot be opened, logs will be written to `stderr`.