Update: Reorganize code. Add new flag(s), implement test flag
Reorganize code, rename `internals/root.go` to be
`internals/rootCmd.go` - add tests for the same. Implement
the test flag with a handler function, will fire commands for
ffmpeg and ffprobe in the backend, fetch the output from the
commands, use regex to extract version text from these commands.

Implement test for test flag. Design base layout for testing
the root command in multiple iterations using randomization(s)
to increase entropy and ensure more resilient testing.

Modify UserInput structure, switch to functional design paradigm
instead of OOP - remove setter method(s). Replace the prior
functionality using `StringVar` and `BoolVar` methods in Cobra
that use a pointer to update the variables directly.

Move validation for user input to inside main method of the root
command instead of structure setter. Add a simple `initialize`
method to the structure to have structure internally process a
part of user input.

Modify `fetchLocation` function to suit a functional approach.
Instead of having the method return errors and set values to
global variables, the function will now return strings containing
the result. In case an error occurs, it will be logged and the
corresponding result will be a blank string for the entry.

Modify tests in accordance with the changes made so far.

Modidy github linter to not run `GoSec` - it is a part of
GolangCI-Lint, remove entry for the same from Github workers,
Makefile and the setup bash script as well.