Update: Major changes. Initial usable version
The largest update with this commit being the fact that this commit
happens to present the first usable version of the application.

Is the app stable enough for a release yet? No. Buggy as it may be
at the moment, this version is still usable for the most part.

What's remaining? A lot of minor household tasks; having the
application clean up after itself, better logging, modifying
behaviour based on the minor flags (`--echo` for one), improving
the overall codebase and a bit of other fluff.

As for the changes in this commit, in no particular order;

 - Check the path to the root directory, force-stop if the path
is incorrect, or points to a file

 - Iterate through list of all directories present inside the
root directory. Sort files present in these source directories
into categories based on file extensions.

 - File categories allowed; media file, subtitles, attachments,
and chapters. Each category allowing a unique set of file
extensions, and requiring different metadata/input format for the
final FFmpeg command being formed.

 - Based on this, throw an error if no media file can be detected,
or if no subtitle/attachment/chapters files can be found. Ignore
files with unrecognized extensions.

 - Form the final FFmpeg command based on file categories, create
a directory named `results` in the root directory and place output
files from FFmpeg inside this directory.

 - Rename and move `internals/userInput.go` to now be
`internals/commons/commons.go`. This change allows common resources
required by `rootCmd.go` and `handler.go` to be accessed freely
without a chain of cyclic dependencies.

 - Modify logging statements to point to the file/method they
originate from; allowing easy traceback to the point at which the
log message was fired.

 - Minor changes to the codebase to improve documentation, and
readibility.

 - Modify test functions to avoid a deferred call to
`monkey.UnpatchAll()`; instead replacing it with deferred calls
for individual patches.

 - Minor bug patches, change output statements.