BugFix: Patch a minor bug. Add new tests. Refactor
Patch a minor bug caused by the test flag - the flag attempted
to perform a call to the executables without validating the paths
before.

Remove the check to validate the root path - instead, club the check
with `ioutil.ReadDir` - the function call will simply return an error
if the root path is invalid, or does not point to a directory.

Add tests for all changes made in this commit, improve testing,
cover edge-cases, and improve documentation for tests.

Convert the list of exit codes to be constant - as it should have been.
Using Python for too long made me forget that 'normal' languages have
constants :p