Update: Add new tests, populate testdata with samples
Add new tests centered around testing the functionality of
`handler.go` and its internal working.

Modify the internal structure of the application. Add a helper
method to `UserInput` that will internally perform a check to
see if a file is to be skipped (because of regex exclusion, or
exclusion list).

Other minor bug-fixing.

Remove the `echo` field from user input and its corresponding flag.
The flag might be re-introduced in the future, but as of now, it
just feels out-of-the place.

Modify the Github workflow for CI to include the `gcflags` flag
required to monkey patch - was resulting in failing test cases,
took a long time to locate :(

Reduce minimum coverage threshold for Codecov momentarily.