BugFix: Make files be case-insensitive, ignore directories
File extensions were not being checked for case insensitivity,
i.e. a file `subtitle.VTT` was not being recognized as a valid
subtitle file, this commit patches this up.

Ignore (sub)directories while grouping all the files present in
a source directory - not really a bug, but lead to annoying log
messages that a directory is being ignored cause it couldn't
be recognized by its extension (duh!)

Minor changes to the tests, also, reducing the relative coverage
required to get CodeCov to pass tests. Will be increasing this
once the project is ready for a release!