Major: Reimplement update queue. Patch bugs. Refactor
Re-implement the logic to fetch and display real time updates
on the screen. This commit separates the code for updates from
the rest of the application (unlike the previous commit, which
dumped it all together :/)

Create a simple structure to contain all the code related to
displaying updates in the background. Also, implement a progress
bar for graphics along with progress percentage counter. Convert
file size from raw numbers into a read-able version.

Patch bug that caused the previous update queue to stop updates
midway, the bug was caused when the main thread flipped the kill
switch for the goroutine as soon as the running command would
complete - causing the thread providing updates to stop midway

The re-implemented logic overcomes the bug above, as well as
updates the progress bar, file size, etc values when the command
fired completes. Replace the one-way channel from the main
thread to the goroutine  with a two-way channel.

Once the background command completes execution, the main thread
will fire a signal on the channel, when the goroutine receives
this signal it will perform a final update on the screen, and
then fire another signal back as an acknowledgement.

Remove all the previous logic regarding the background thread, it
had its own fair share of bugs, it was easier to re-write a new
logic than attempting to patch the previous one :D

Change a few log statements for better readability. Cover a large
part of the codebase under unit-tests.