Minor: Refactor. Polish the codebase. Modify github actions
Minor refactoring. Remove the `getRootCommand` function being
used as a generator to create new instances of root command on
the fly.

Replace it with a central copy of the root command (because more
than one copy isn't required). Tests can create a copy of this
variable when required.

Additional minor changes to the code-base. Trim unnecessary
comments, attempt to better explain changes being made. Remove
peices of code that weren't required, minor changes to the help
text.

Modify Github actions to prevent triggering them for brances
starting with `rt-**` or `docs-**`. Both of these naming patterns
will be reserved. With `rt` being the rapid-testing branch which
shall have frequent updates, at the cost of breaking tests,
untested changes and more. The docs branch will be used for
minor (and major) changes being made to the documentation.