Merge rapid-testing 0.0.1 (#1)
* Minor: Modify github actions, add tests

Minor refactoring. Remove the `getRootCommand` function being
used as a generator to create new instances of root command on
the fly.

Replace it with a central copy of the root command (because more
than one copy isn't required). Tests can create a copy of this
variable when required.

Additional minor changes to the code-base. Trim unnecessary
comments, attempt to better explain changes being made. Remove
pieces of code that weren't required, minor changes to the help
text.

Modify Github actions to prevent triggering them for branches
starting with `rt-**` or `docs-**`. Both of these naming patterns
will be reserved. With `rt` being the rapid-testing branch which
shall have frequent updates, at the cost of breaking tests,
untested changes and more. The docs branch will be used for
minor (and major) changes being made to the documentation.

* Minor: Update log messages, fix code documentation

Update/add log messages to reflect the changes made in master
branch so far, add new logging statement(s) to track the
flow-of-control better (to simply bug-reports).

Fix code documentation, update redundant comments from a long
time ago.

Define a new function in `internals/rootCmd.go` to separately
handle firing the backend functions in case the `--test` flag
is being used. Change the flow-of-control to throw an error
on screen if the `ffmpeg/TraverseRoot` function fails
internally.

* Update: Add help string, increase test coverage, minor refactor

Add help string explaining the basic usage of the command.

Add new tests to increase the overall test-coverage. Remove
a few redundant tests that are no longer required - the
functionality they were testing was either removed, or covered
with better test(s)