Feature: Add force-flag. Minor refactor
Add the force-flag, can be enabled with the shorthand notification
of `-f`, or using `--force`. If the force-flag is enabled, auto-sub
will append all ffmpeg commands with the `-f` flag to ensure
existing files are overwritten if required.

Minor bug-patch, create a separate method under `UserInput` structure
to generate the final name of the output file - this change will be
required in a future commit.